Refactor: Enhance Orders Feature
This commit introduces several improvements to the Orders feature, including UI refactoring, code cleanup, and enhanced error handling.

Key changes:
- Refactored `OrdersScreen` by extracting the orders display logic into a new widget `ShowOrders`. This improves code organization and readability.
- Removed unnecessary imports and newlines across various files in the Orders feature, including DTOs, mappers, services, and repositories.
- Renamed `OrdersService` to `OrdersApiService` for better clarity and consistency.
- Updated `OrdersRemoteDataSourceImpl` to utilize the `executeApi` helper function for streamlined API call handling and error management.
- Improved null safety in `MetadataMapper` and `OrderResponseMapper` by providing default values for nullable fields, ensuring more robust data mapping.
- Adjusted `OrdersState` by removing the `clearFailure` and `clearOrders` parameters from the `copyWith` method as they are no longer needed.
- Removed the unused `discount` field from `ProductDetailsModel`.
