+++ lib/Feature/ordersPage/api/models/metadata_dto.dart
@@ -1,5 +1,4 @@
 import 'package:json_annotation/json_annotation.dart';
-
 part 'metadata_dto.g.dart';
 
 @JsonSerializable()
