+++ lib/Feature/ordersPage/api/models/product_dto.dart
@@ -1,5 +1,4 @@
 import 'package:json_annotation/json_annotation.dart';
-
 part 'product_dto.g.dart';
 
 @JsonSerializable()
